Fixed warnings, fixed hex-editor ruler.

+++ Tools/LSWindows/Src/TreeListView/LSWTreeListView.cpp
@@ -613,10 +613,10 @@ namespace lsw {
 	BOOL CTreeListView::SortItems( INT _iSubItem ) {
 		if ( !m_bSort ) { return FALSE; }
 		try {
-			if ( _iSubItem >= m_tRoot.Value().vSortAscNext.size() ) {
+			if ( _iSubItem >= INT( m_tRoot.Value().vSortAscNext.size() ) ) {
 				m_tRoot.Value().vSortAscNext.resize( _iSubItem + 1 );
 			}
-			auto eSortMethod = (_iSubItem < m_vSortMethod.size()) ? m_vSortMethod[_iSubItem] : ee::EE_TP_ORDINAL_CI;
+			auto eSortMethod = (_iSubItem < INT( m_vSortMethod.size() )) ? m_vSortMethod[_iSubItem] : ee::EE_TP_ORDINAL_CI;
 			SortChildren( &m_tRoot, _iSubItem, m_tRoot.Value().vSortAscNext[_iSubItem], eSortMethod );
 			m_tRoot.Value().vSortAscNext[_iSubItem] = !m_tRoot.Value().vSortAscNext[_iSubItem];
 			UpdateListView();
@@ -796,10 +796,10 @@ namespace lsw {
 	 **/
 	void CTreeListView::DeleteColumn( ee::CTree<CTreeListView::LSW_TREE_ROW> * _ptThis, INT _iCol ) {
 		if ( _ptThis ) {
-			if ( _iCol < _ptThis->Value().vSortAscNext.size() ) {
+			if ( _iCol < INT( _ptThis->Value().vSortAscNext.size() ) ) {
 				_ptThis->Value().vSortAscNext.erase( _ptThis->Value().vSortAscNext.begin() + _iCol );
 			}
-			if ( _iCol < _ptThis->Value().vStrings.size() ) {
+			if ( _iCol < INT( _ptThis->Value().vStrings.size() ) ) {
 				_ptThis->Value().vStrings.erase( _ptThis->Value().vStrings.begin() + _iCol );
 			}
 			for ( auto I = _ptThis->Size(); I--; ) {
